+++ src/rosros/ros2.py
@@ -8,7 +8,7 @@ Released under the BSD License.
 
 @author      Erki Suurjaak
 @created     11.02.2022
-@modified    28.02.2023
+@modified    03.12.2023
 ------------------------------------------------------------------------------
 """
 ## @namespace rosros.ros2
@@ -1281,13 +1281,17 @@ def _get_message_type_hash(typename):
     return parsing.calculate_definition_hash(typename, definition)
 
 
-def get_message_value(msg, name):
+def get_message_value(msg, name, default=...):
     """
     Returns object attribute value, with numeric arrays converted to lists.
 
-    @param   message attribute name; may also be (nested, path) or "nested.path"
+    @param   name     message attribute name; may also be (nested, path) or "nested.path"
+    @param   default  value to return if attribute does not exist; raises exception otherwise
     """
-    v, parent, k = util.get_nested(msg, name)
+    try: v, parent, k = util.get_nested(msg, name)
+    except Exception:
+        if default is not Ellipsis: return default
+        raise
     if isinstance(v, (bytes, array.array)) \
     or "numpy.ndarray" == "%s.%s" % (v.__class__.__module__, v.__class__.__name__):
         v = list(v)