+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog for package rosros
 
 0.2.1 (2023-12-03)
 ------------------
+* add default fallback parameter to api.get_message_value()
 * fix util.get_arity() counting some parameters twice if counting both positional and keyword arguments
 * fix api.time_message() not using clock_type parameter in ROS2